[llm] Move OpenAILLm to its own file
This better conforms with our code base styling
