+++ src/canopy/llm/__init__.py
@@ -1,3 +1,4 @@
 from .base import BaseLLM
 from .openai import OpenAILLM
 from .anyscale import AnyscaleLLM
+from .azure_openai_llm import AzureOpenAILLM
